Connected the backend wtih frontend using websocket, created additinal security config and cors config in backend

+++ Server/src/main/java/lk/ac/iit/Mihin/Server/Repositories/ConfigurationRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConfigurationRepository extends JpaRepository<Configuration, Long> {
+public interface ConfigurationRepository extends JpaRepository<Configuration, Integer> {
+    // Additional query methods if needed
 }
