+++ Server/src/main/java/lk/ac/iit/Mihin/Server/Repositories/VendorRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VendorRepository extends JpaRepository<Vendor, Long> {
-    // Custom query methods (if needed)
+public interface VendorRepository extends JpaRepository<Vendor, Integer> {
+    // Additional query methods if needed
 }
+
